stockCategory and deliveryNotes Changes Fixes

+++ src/main/java/com/asg/shipchandling/stockcategory/dto/StockCategoryGlValuesDto.java
@@ -5,8 +5,6 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.math.BigDecimal;
-
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
@@ -14,7 +12,7 @@ public class StockCategoryGlValuesDto {
     private Long parentCategoryPoid;
     private String parentCategoryCode;
     private String parentCategoryName;
-    private BigDecimal stockGlPoid;
-    private BigDecimal salesGlPoid;
-    private BigDecimal costOfSalesGlPoid;
+    private Long stockGlPoid;
+    private Long salesGlPoid;
+    private Long costOfSalesGlPoid;
 }
